// src/api/generated/models/auditLogsControllerFindAllParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { PrismaAuditAction } from './prismaAuditAction';

export type AuditLogsControllerFindAllParams = {
entityType?: string;
entityId?: string;
/**
 * Prisma audit action: create | update | delete
 */
action?: PrismaAuditAction;
/**
 * Filter by changedBy user id
 */
userId?: string;
/**
 * Inclusive start datetime filter
 */
from?: string;
/**
 * Inclusive end datetime filter
 */
to?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};