// src/api/generated/models/paginatedAuditLogsResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */
import type { AuditLogResponseDto } from './auditLogResponseDto';

export interface PaginatedAuditLogsResponseDto {
  items: AuditLogResponseDto[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}